refactor(editor): extraire EditorToolbar (mutualise normal + plein écran) (#28)
Tranche 2 de la décomposition de TipTapEditor.vue.

- EditorToolbar.vue (<script setup>, fragment) : barre d'outils partagée
  (font/format/couleur/alignement/titres/listes/table/blocs/média/undo-redo).
  Props editor + showFullscreenToggle ; émet toggle-fullscreen. Handlers à
  prompt (table/lien/image/vidéo) locaux. CSS des boutons copiée à l'identique.
- TipTapEditor : les DEUX toolbars (mode normal + mode plein écran), aux jeux
  de commandes vérifiés identiques, utilisent désormais <EditorToolbar>. Markup
  dupliqué (~630 l.) + CSS boutons + handlers morts retirés (1611 -> 859 l.).
  addLink conservé (bubble menu). Conteneurs .editor-toolbar-modern /
  .modal-toolbar conservés au parent.

Cumul TipTapEditor : 1679 -> 859 l. (-49 %).

Parité visuelle : CSS à l'identique + build vert, QA manuelle recommandée
(toolbar en mode normal ET plein écran).

Tests 261/261, contract 28/28, build OK.
